pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = [{ name = "Seu nome aqui", email = "seu.email@email.com.br" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "ipdb (>=0.13.13,<0.14.0)",
    "celery[redis] (>=5.4.0,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    "channels (>=4.0.0,<5.0.0)",
    "channels-redis (>=4.2.0,<5.0.0)",
    "daphne (>=4.0.0,<5.0.0)",
    "django-cors-headers (>=4.3.0,<5.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
